Globalized VidResultContainers
